fix: fix small bags
