+++ roulette/models/completeModel.js
@@ -1,5 +1,3 @@
-// import { minBet } from './consts/min_max.js';
-// import { maxBet } from './consts/min_max.js';
 import { payoutRatios } from './consts/payoutRatios.js';
 import { rouletteNumber } from './consts/rouletteNumber.js';
 // import { rltPos } from './consts/rltPos.js';
@@ -18,11 +16,13 @@ calculate.onclick = function () {
 
     // блок функций:
     function chipsNeededForABet(strUp) {
+        // повторяющийся код:
         let quantityNumb = payoutRatios.numb.position * rouletteNumber[strUp].numb;
         let quantitySplit = payoutRatios.split.position * rouletteNumber[strUp].split;
         let quantityCorner = payoutRatios.corner.position * rouletteNumber[strUp].corner;
         let quantityStreet = payoutRatios.street.position * rouletteNumber[strUp].street;
         let quantitySix_line = payoutRatios.six_line.position * rouletteNumber[strUp].six_line;
+        //
         let quantitySum = quantityNumb + quantitySplit + quantityCorner + quantityStreet + quantitySix_line;
         let result = quantitySum;
         return result;
