+++ roulette/models/fieldAndTrackModel.js
@@ -1,18 +1,11 @@
-// import "./menuModel.js";
 // import { minBet } from "./menuModel.js";
 // import { maxBet } from "./menuModel.js";
 // import { payoutRatios } from "./consts/payoutRatios.js";
-// import { chipsNeededForABet } from "./menuModel.js";
-// import { countPositions } from "./menuModel.js";
 // import { oddDozColum } from "./consts/oddDozColum.js";
 // import { rouletteSeries } from './consts/rouletteSeries.js';
-
-import { rouletteNumber } from "./consts/rouletteNumber.js";
-import { fullBets } from "./functions/fullBets.js";
-import { rltPos } from "./consts/rltPos.js";
-
-// поворот экрана:
-// document.body.style.transform = 'rotate(90deg)';
+// import { rouletteNumber } from "./consts/rouletteNumber.js";
+// import { fullBets } from "./functions/fullBets.js";
+// import { rltPos } from "./consts/rltPos.js";
 
 // пересекается ли трек с полем:
 export const trackOverField = true;
@@ -25,7 +18,7 @@ nodeList.forEach(element => {
         let text = element.textContent;
         // Выводим текст элемента в консоль и всплывашку:
         console.log(text);
-        alert(text);
+        // alert(text);
     });
 });
