+++ roulette/models/seriesModel.js
@@ -1,4 +1,3 @@
-// import "./menuModel.js";
 import { rouletteSeries } from "./consts/rouletteSeries.js";
 import { payoutRatios } from "./consts/payoutRatios.js";
