+++ roulette/views/menu/menuView.html
@@ -53,16 +53,8 @@
             </ul>
         </nav>
     </div>
-    <!-- невидимая кнопка -->
-    <div>
-        <div class="btn">
-            <input id="calculate" type="button" value="Рассчитать" />
-        </div>
-    </div>
 
-    <script src="../../models/consts/min_max.js" type="module"></script>
-    <script src="../../models/consts/payoutRatios.js" type="module"></script>
-    <script src="../../models/consts/rouletteNumber.js" type="module"></script>
+    <!-- <script src="../../models/consts/min_max.js" type="module"></script> -->
     <script src="../../models/menuModel.js" type="module"></script>
 </body>
